sets up import job and notification

+++ app/Imports/ScreeningImporter.php
@@ -2,13 +2,11 @@
 
 namespace App\Imports;
 
-use App\Models\Dimension;
-use App\Models\Geography;
-use App\Models\Metric;
+
 use App\Models\Reference;
-use App\Models\Scale;
+use App\Jobs\ProcessImporter;
 use Illuminate\Support\Collection;
-use Illuminate\Support\Str;
+use App\Jobs\ProcessImportNotifcation;
 use Maatwebsite\Excel\Concerns\ToCollection;
 use Maatwebsite\Excel\Concerns\WithHeadingRow;
 
@@ -24,77 +22,24 @@ class ScreeningImporter implements ToCollection, WithHeadingRow
 
             $reference = Reference::firstWhere('rayyan_key', $row['key']);
 
-            if($row['decision'] === 'exclude') {
+            if($row['decision'] === 'exclude' && isset($reference)) {
 
                 $reference->delete();
 
             }
 
-            elseif($row['decision'] === 'include') {
+            elseif($row['decision'] === 'include' && is_null($reference->imported_entities)) {
 
-                $str = Str::of($row['corrected'])
-                    ->split('/[\n\r]/');
+                $reference->update(['imported_entities' => $row['corrected']]);
 
-                foreach($str as $line) {
-
-                    $line = Str::of($line)
-                        ->trim();
-
-                    $type = $line->before(":");
-                    $items = $line->after(':');
-
-                    $items = collect(str_getcsv($items));
-
-                    $items = $items->map(fn($item) => Str::of($item)
-                        ->trim()
-                        ->toString())
-                        ->filter(fn($item) => $item !== '');
-
-                    foreach ($items as $item) {
-
-                        if($type->toString() === "Metrics") {
-
-                            $metric = Metric::updateOrCreate(['title' => $item]);
-                            $reference->metrics()->syncWithoutDetaching([$metric->id => ['reference_type' => 'reference']]);
-
-                        }
-
-                        elseif($type->toString()=== "Dimensions") {
-
-                            $dimension = Dimension::updateOrCreate(['name' => $item]);
-                            $reference->dimensions()->syncWithoutDetaching([$dimension->id => ['reference_type' => 'reference']]);
-
-                        }
-
-//                        elseif($type->toString()=== "Methods") {
-//
-//
-////                            $method = Method::updateOrCreate(['name' => $item]);
-////                            $reference->methods()->syncWithoutDetaching([$method->id => ['reference_type' => 'reference']]);
-//
-//                        }
-
-                        elseif($type->toString()=== "Country/region") {
-
-                            $geography = Geography::updateOrCreate(['name' => $item]);
-                            $reference->geographies()->syncWithoutDetaching([$geography->id => ['reference_type' => 'reference']]);
-
-                        }
-
-                        elseif($type->toString()=== "Scale" || $type->toString()===  "Level") {
-
-                            $scale = Scale::updateOrCreate(['name' => $item]);
-                            $reference->scales()->syncWithoutDetaching([$scale->id => ['reference_type' => 'reference']]);
-
-                        }
-
-                    }
-
-                }
+                ProcessImporter::dispatch($reference);
 
             }
 
         }
+
+        $recipient = auth()->user();
+        ProcessImportNotifcation::dispatch($recipient);
 
     }
